+++ src/features/Character/components/Skills/index.js
@@ -6,9 +6,9 @@ import TooltipTrigger from 'common/components/TooltipTrigger';
 import Icon from 'common/components/Icon';
 import styles from './styles.less';
 
-const Skill = ({ data = {}, className }) => (
-  <TooltipTrigger type="skill" data={data}>
-    <Icon src={data.icon} size="medium" className={cx(styles.skill, className)} />
+const Skill = ({ data, className }) => (
+  <TooltipTrigger type="skill" data={data || 'No Skill Selected'}>
+    <Icon src={get(data, 'icon')} size="medium" className={cx(styles.skill, className)} />
   </TooltipTrigger>
 );